python_engine/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
requests==2.31.0
lightgbm==4.1.0
